AI Agent Hackathon Code/Linkedin_AI_Agent/requirements.txt
python-dotenv==1.0.0
selenium==4.18.1
groq==0.4.2
linkedin-api==2.0.0
webdriver_manager==4.0.1
pydantic==1.10.13
